requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
whitenoise==6.8.2
requests==2.32.3